Add development.json.sample
