+++ config/.gitignore
@@ -1,2 +1,3 @@
 *.json
 !defaults.json
+!*.sample
