--- lib/config/.gitignore
@@ -1,1 +1,0 @@
-client.js
