+++ lib/config/cast-string.js
