--- lib/config/config.js
@@ -1,7 +1,0 @@
-/**
- * Module dependencies.
- */
-
-var client = require('./client');
-
-module.exports = window.config ? window.config : client;
