+++ lib/config/index.js
@@ -1,48 +1,90 @@
 /**
- * Extend module's NODE_PATH
- * HACK: temporary solution
- */
-
-require('node-path')(module);
-
-/**
- * Module dependencies.
- */
+* Configuration
+*
+* - Default config options will be loaded from `/config/defaults.json`. Also it
+*   will be used as reference for the Type the values should have.
+*
+* - Environment specific overrides are optional, using `/config/{NODE_ENV}.json`.
+*
+* - Environment Variables also can be used to override options (recommended for
+*   production).
+*   + Var names should be CONSTANT_CASE.
+*     + e.g.: `mongoUrl` => `MONGO_URL`
+*     + Scoped variables e.g.: `notifications.url` => `NOTIFICATIONS_URL`
+*   + `Arrays`s should be strings separated by commas.
+*     + e.g.: `"staff": ["mail@eg.com", "a@c.m"]` => `STAFF="mail@eg.com,a@c.m"`
+*   + `Boolean`s should be `true` or `false` as strings.
+*     + e.g.: `"rssEnabled": false` => `RSS_ENABLED="false"`
+*   + Var Types are casted using `./cast-string.js`
+**/
 
 var log = require('debug')('democracyos:config');
-var environment = process.env.NODE_ENV || 'development';
 var resolve = require('path').resolve;
-var merge = require('merge-util');
-var utils = require('lib/utils');
-var has = utils.has;
-var md5 = utils.md5;
-var envConf = require('./env');
+var path = require('path');
+var typeOf = require('component-type');
+var changeCase = require('change-case');
+var cast = require('./cast-string');
 
-var filepath = resolve(__dirname, "..", "..", "config", environment + ".json");
+var env = process.env;
+var environment = env.NODE_ENV || 'development';
 
-var localConf = {};
+var configPath = resolve(__dirname, '..', '..', 'config');
+var defaultsPath = resolve(configPath, 'defaults.json');
+var envPath = resolve(configPath, environment + '.json');
 
-try {
-  log('Load local configuration from %s', filepath);
-  localConf = require(filepath);
-} catch (e) {
-  log('Unalbe to read configuration from file %s: %s', filepath, e.message);
+var defaultConfig = require(defaultsPath);
+var localConfig = fs.existsSync(envPath) && require(envPath) || {};
+var config = {};
+
+forEach(defaultConfig, parse);
+
+function parse(val, key, scope){
+  var s = scope ? scope.slice(0) : [];
+  var c = get(config, s);
+
+  if (typeOf(val) === 'object') {
+    c[key] = {};
+    forEach(val, parse, s.concat(key));
+    return;
+  }
+
+  var envKey = s.concat(key).map(changeCase.constantCase).join('_');
+  if (env.hasOwnProperty(envKey)) {
+    var newVal;
+    try {
+      newVal = cast(typeOf(val), env[envKey]);
+    } catch(e) {
+      throw new Error('There was an error when parsing ENV "' + envKey + '": ' + e);
+    }
+    return c[key] = newVal;
+  }
+
+  var local = get(localConfig, s);
+  if (local.hasOwnProperty(key)) {
+    var newVal = local[key];
+    if (typeOf(val) !== typeOf(newVal)) {
+      throw new Error('Invalid value for key "' + key + '" on "' + environment + '.json": ' + '". Should be "' + typeOf(val) + '".');
+    }
+    return c[key] = newVal;
+  }
+
+  c[key] = val;
 }
 
-log('Merge environment set configuration variables');
+function get(obj, scope) {
+  var c = obj;
+  if (scope) scope.forEach(function(k){ c = c[k] });
+  return c;
+}
 
-var conf = merge(localConf, envConf);
-conf.env = environment;
-
-log('Loaded config object for env %s with sha %j', environment, md5(JSON.stringify(conf)));
+function forEach(obj, cb) {
+  var extraArgs = [].slice.call(arguments, 2);
+  Object.keys(obj).forEach(function(key) {
+    var val = obj[key];
+    var args = [val, key];
+    if (extraArgs.length) args = args.concat(extraArgs);
+    cb.apply(obj, args);
+  });
+}
 
 module.exports = config;
-
-function config(key) {
-  if (has.call(conf, key)) return conf[key];
-  log('Invalid config key "%s"', key);
-  // throw new Error('Invalid config key "' + key + '"');
-  return undefined;
-}
-
-for (var key in conf) config[key] = conf[key];
